test: Validator 테스트 코드 ValidatorTest로 분리
